Apply versioning to the api
