+++ app/controllers/api.rb
@@ -1,0 +1,2 @@
+module Api
+end
