+++ app/controllers/api/v0.rb
@@ -1,0 +1,4 @@
+module Api
+  module V0
+  end
+end
