+++ app/controllers/api/v0/users_controller.rb
@@ -8,5 +8,9 @@ Contact: you@your-company.com
 Generated by: https://github.com/swagger-api/swagger-codegen.git
 
 =end
-class UsersController < BaseController
+module Api
+  module V0
+    class UsersController < BaseController
+    end
+  end
 end
