+++ app/controllers/api/v0x0.rb
@@ -1,0 +1,7 @@
+module Api
+  module V0x0
+    class AdminsController < Api::V0::AdminsController; end
+    class UsersController < Api::V0::UsersController; end
+    class BaseController < Api::V0::BaseController; end
+  end
+end
