+++ config/routes.rb
@@ -10,16 +10,19 @@ Generated by: https://github.com/swagger-api/swagger-codegen.git
 =end
 Rails.application.routes.draw do
 
-  mount SwaggerUiEngine::Engine, at: '/api'
+  mount SwaggerUiEngine::Engine, at: '/open-api'
 
   def add_swagger_route http_method, path, opts = {}
     # Convert codegen substitutions into Symbols
     full_path = path.gsub(/{(.*?)}/, ':\1')
     # Add back in a custom base path if required
-    full_path = File.join(ENV["BASE_PATH"], full_path) if ENV["BASE_PATH"]
-
-    constraint = opts[:constraint_name].camelize.constantize
-    match full_path, to: "#{opts.fetch(:controller_name)}##{opts[:action_name]}", constraints: constraint, via: http_method
+    namespace :api do
+      namespace :v0x0, :path => "v0.0" do
+        full_path = File.join(ENV["BASE_PATH"], full_path) if ENV["BASE_PATH"]
+        constraint = opts[:constraint_name].camelize.constantize
+        match full_path, to: "#{opts.fetch(:controller_name)}##{opts[:action_name]}", constraints: constraint, via: http_method
+      end
+    end
   end
 
   add_swagger_route 'POST', '/portfolios', controller_name: 'admins', constraint_name: 'adminsConstraint', action_name: 'add_portfolio'
